+++ doc/changelog.md
@@ -1,12 +1,13 @@
 # 2020년 1월 3일
 
 - 프로젝트 생성
-- golf, windmill 앱 생성
+- golf, windmill, member 앱 생성
+- local, production 환경 분리
+- secret.json
+- static / media files
 
 # TODO
-- local, production 환경 분리
 - 템플릿 환경 만들기
-- static assets / upload
 - django-auth 연동 (가입 및 절차)
 - 라인/카톡 연동
 - 역할 그룹 분류: 골프장, 여행사(회사/개인), 나원, 고객